Запрет удаления работника, если он является чьим-то руководителем

+++ TelSpr/MainForm.cs
@@ -976,10 +976,31 @@ namespace TelSpr
                 return;
             }
 
-            var parameters = new Dictionary<string, object> {{"id", currRow["id"]}};
-            DBFunctions.ExecuteCommand("DELETE FROM workers WHERE id=@id",parameters);
 
-            LoadWorkers();
+            var parameters = new Dictionary<string, object> { { "id", currRow["id"] } };
+            DBFunctions.ExecuteCommand("DELETE FROM workers WHERE id=@id", parameters);
+
+            var dtChildWorkers = DBFunctions.ReadFromDB(@"SELECT 
+            workers.id as id, 
+            workers.second_name || ' ' || workers.name || 
+            CASE WHEN workers.third_name NOT NULL AND workers.third_name <> '' THEN ' ' || workers.third_name ELSE '' END AS FIO
+            FROM workers WHERE boss = @id",parameters);
+
+            if(dtChildWorkers.Rows.Count == 0)
+            {
+                DBFunctions.ExecuteCommand("DELETE FROM workers WHERE id=@id", parameters);
+                LoadWorkers();
+            } else
+            {
+                string messageText = "Удаление невозможно. Сотрудник является руководителем для: ";
+
+                foreach(DataRow childWorkerRow in dtChildWorkers.Rows)
+                {
+                    messageText += Environment.NewLine + childWorkerRow["FIO"] + " (" + childWorkerRow["id"] + ")";
+                }
+
+                MessageBox.Show(messageText);
+            }           
 
         }
     }
